backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
pydantic==2.10.6
python-dotenv==1.0.1
sse-starlette==2.2.1
httpx==0.28.1
